Import-OOXML Improvement
The Import-OOXML always consider that first data row is actualy de data
header that will be used to generate the data object model. In the
previous versions, there was no check to see if a header was empty or
null. Now the Cmdlet will read each header and will stop if no header
value is found.